Bugs with declatations in convert.h and functions fixed. Bugs with somes functions fixed too.

+++ includes/convert.h
@@ -16,14 +16,14 @@
 # include "libft.h"
 
 int		ft_atoi_len(const char *str, int *len);
-char	*ft_ultoa_to(int n, char *dest);
-char	*ft_uitoa_to(int n, char *dest);
+char	*ft_ultoa_to(unsigned long n, char *dest);
+char	*ft_uitoa_to(unsigned int n, char *dest);
 char	*ft_ltoa_to(long n, char *dest);
 char	*ft_itoa_to(int n, char *dest);
 int		ft_atoi(const char *str);
-char	*ft_ultoa(int n);
-char	*ft_uitoa(int n);
-char	*ft_ltoa(int n);
+char	*ft_ultoa(unsigned long n);
+char	*ft_uitoa(unsigned int n);
+char	*ft_ltoa(long n);
 char	*ft_itoa(int n);
 
 #endif
